+++ CHANGELOG.md
@@ -10,6 +10,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - Added the ability to do simple copies, no need to care about `copy/{tag,type,all}` anymore if you don't want to.
 - Added support for naming a repo different than the folder name.  Use `name` in opts.yml.
 - Added a User-Agent so Docker and other stuff can identify our software.
+- Complex aliased metadata is not treated like a full build. [#7]
 - Added a default timeout of 480 to Excon.
 
 #### Changed